Add new interface to move the code out from PersistenceInterface. Fix review comments and typos
